+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.6] - Network Simulation & Dependency Updates
+* [new] network Speed Throttling (Slow 3G, Offline Mode, Socket/Server Exceptions)
+* [optimization] log entry limit (30 items)
+* [update] floating logger control show max log & network simulation
+
 ## [0.1.5] - Runtime and fixing formdata request
 * [fix] add runtime to logger
 * [fix] fixing formdata request